add and test player movement sync

+++ Assets/_Scripts/Network/NetworkMessages/NetworkMessage.cs
@@ -18,7 +18,7 @@ public abstract class ClientMessage : NetworkMessage
     [JsonProperty("clientId")]
     public string SenderId;
     [JsonProperty("clientSendTime")]
-    public double ClientSendTime;
+    public long ClientSendTime;
     [JsonProperty("clientTick")]
     public int ClientTick;
     protected ClientMessage(short messageType) : base(messageType) { ClientSendTime = TimeSyncUtils.GetUnixTimeMilliseconds(); }
@@ -29,7 +29,7 @@ public abstract class ServerMessage : NetworkMessage
     [JsonProperty("clientId")]
     public string ReceiverId;
     [JsonProperty("serverSendTime")]
-    public double ServerSendTime;
+    public long ServerSendTime;
     [JsonProperty("serverTick")]
     public int ServerTick;
     protected ServerMessage(short messageType) : base(messageType) { }
